+++ test/test_parseutils.py
@@ -1,6 +1,6 @@
 import pytest
 from mycli.packages.parseutils import (
-    extract_tables, query_starts_with, queries_start_with, is_destructive
+    extract_tables, query_starts_with, queries_start_with, is_destructive, query_has_where_clause
 )
 
 
@@ -135,3 +135,32 @@ def test_is_destructive():
         'drop database foo;'
     )
     assert is_destructive(sql) is True
+
+
+def test_is_destructive_update_with_where_clause():
+    sql = (
+        'use test;\n'
+        'show databases;\n'
+        'UPDATE test SET x = 1 WHERE id = 1;'
+    )
+    assert is_destructive(sql) is False
+
+
+def test_is_destructive_update_without_where_clause():
+    sql = (
+        'use test;\n'
+        'show databases;\n'
+        'UPDATE test SET x = 1;'
+    )
+    assert is_destructive(sql) is True
+
+
+@pytest.mark.parametrize(
+    ('sql', 'has_where_clause'),
+    [
+        ('update test set dummy = 1;', False),
+        ('update test set dummy = 1 where id = 1);', True),
+    ],
+)
+def test_query_has_where_clause(sql, has_where_clause):
+    assert query_has_where_clause(sql) is has_where_clause